+++ test/benchmark_tall_stairs.py
@@ -60,6 +60,7 @@ def make_planner(m: Map2D5, disable_clearance: bool) -> HoppingAStarPlanner:
         robot_radius=config.ROBOT_RADIUS,
         leg_length=config.LEG_LENGTH,
         min_clearance_gate=config.MIN_CLEARANCE,
+        steep_grade=config.STEEP_INFLATE_GRADE,
         arc_max_step=config.ARC_SAMPLE_MAX_STEP,
         hop_scan_step=config.HOP_SCAN_STEP,
         hop_scan_step_ref_radius=config.HOP_SCAN_STEP_REF_RADIUS,