+++ test/demo_planner_reroute.py
@@ -103,6 +103,7 @@ def make_planner(m: Map2D5) -> HoppingAStarPlanner:
         robot_radius=config.ROBOT_RADIUS,
         leg_length=config.LEG_LENGTH,
         min_clearance_gate=config.MIN_CLEARANCE,
+        steep_grade=config.STEEP_INFLATE_GRADE,
         arc_max_step=config.ARC_SAMPLE_MAX_STEP,
         hop_scan_step=config.HOP_SCAN_STEP,
         hop_scan_step_ref_radius=config.HOP_SCAN_STEP_REF_RADIUS,