+++ test/test_hop_energy_chain.py
@@ -80,6 +80,7 @@ def _make_planner(m: Map2D5, start, goal, **overrides) -> HoppingAStarPlanner:
         speed_bin=config.SPEED_BIN, mu=config.MU,
         robot_radius=config.ROBOT_RADIUS, leg_length=config.LEG_LENGTH,
         min_clearance_gate=config.MIN_CLEARANCE,
+        steep_grade=config.STEEP_INFLATE_GRADE,
         arc_max_step=config.ARC_SAMPLE_MAX_STEP,
         hop_scan_step=config.HOP_SCAN_STEP,
         hop_scan_step_ref_radius=config.HOP_SCAN_STEP_REF_RADIUS,